requirements.txt
streamlit
numpy
open3d
transformers
torch
pillow
opencv-python
